requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
google-generativeai>=0.3.0
requests

# Document processing
pdfplumber
python-pptx
polars
openpyxl
pytesseract
Pillow
img2table
python-docx
ebooklib
pyhwp
nbconvert
PyPDF2

# AI and NLP
langchain-huggingface
sentence-transformers
llama-index
llama-index-llms-gemini

# Miscellaneous
aiofiles
pydantic
git+https://github.com/openai/whisper

# Audio processing
google-cloud-texttospeech==2.14.1
pydub